A05_task/pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "Multi-Source Knowledge Base with AI Agent"
authors = [
    {name = "AI Team"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "langchain>=0.0.335",
    "langchain-openai>=0.0.1",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "pymongo>=4.5.0",
    "pypdf>=3.17.0",
    "sentence-transformers>=2.2.2",
    "tiktoken>=0.5.1",
    "chromadb>=0.4.18",
    "faiss-cpu>=1.7.4",
    "lxml>=4.9.3",
    "python-multipart>=0.0.6",
    "loguru>=0.7.2",
    "aiofiles>=24.1",
    "aiohttp>=3.12.13",
    "gradio>=5.0.0",
    "chardet>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100 
